cloud/Practica1/requirements.txt
pyarrow
pandas
google-cloud-bigquery